користувач може голосувати лише один раз
