+++ shop/urls.py
@@ -15,15 +15,13 @@ Including another URLconf
 """
 from django.conf.urls import url, include
 from django.contrib import admin
-
-import profile.views
 from product import views
 
 urlpatterns = [
     url(r'^admin/', admin.site.urls),
     url(r'^$', views.product_list, name='product_list'),
     url(r'^product/', include('product.urls', namespace='product')),
-    url(r'^logout/', profile.views.logout, name='logout'),
-    url(r'^login/', profile.views.login, name='login'),
+    url(r'^logout/', views.logout, name='logout'),
+    url(r'^login/', views.login, name='login'),
     url(r'^profile/', include('profile.urls', namespace='profile')),
 ]
